add ssmt report rollup service and move ssmt logic
