+++ src/app/report-rollup/report-rollup.service.ts
@@ -28,31 +28,26 @@ import { OpportunityPaybackService } from '../treasure-hunt/treasure-hunt-report
 import { PsatReportRollupService } from './psat-report-rollup.service';
 import { PhastReportRollupService } from './phast-report-rollup.service';
 import { FsatReportRollupService } from './fsat-report-rollup.service';
+import { SsmtReportRollupService } from './ssmt-report-rollup.service';
 
 
 @Injectable()
 export class ReportRollupService {
 
   reportAssessments: BehaviorSubject<Array<ReportItem>>;
-  ssmtAssessments: BehaviorSubject<Array<ReportItem>>;
   treasureHuntAssessments: BehaviorSubject<Array<ReportItem>>;
 
   assessmentsArray: Array<ReportItem>;
-  ssmtArray: Array<ReportItem>;
   treasureHuntArray: Array<ReportItem>;
 
 
 
 
-  selectedSsmt: BehaviorSubject<Array<SsmtCompare>>;
-  ssmtResults: BehaviorSubject<Array<SsmtResultsData>>;
-  allSsmtResults: BehaviorSubject<Array<AllSsmtResultsData>>;
 
   allTreasureHuntResults: BehaviorSubject<Array<TreasureHuntResultsData>>;
 
   calcsArray: Array<Calculator>;
   selectedCalcs: BehaviorSubject<Array<Calculator>>;
-  numSsmt: number = 0;
   numTreasureHunt: number = 0;
   showSummaryModal: BehaviorSubject<string>;
   constructor(
@@ -61,14 +56,14 @@ export class ReportRollupService {
     private assessmentDbService: AssessmentDbService,
     private settingsDbService: SettingsDbService,
     private calculatorDbService: CalculatorDbService,
-    private ssmtService: SsmtService,
     private treasureHuntReportService: TreasureHuntReportService,
     private opportunitySummaryService: OpportunitySummaryService,
     private opportunityCardsService: OpportunityCardsService,
     private opportunityPaybackService: OpportunityPaybackService,
     private psatReportRollupService: PsatReportRollupService,
     private phastReportRollupService: PhastReportRollupService,
-    private fsatReportRollupService: FsatReportRollupService
+    private fsatReportRollupService: FsatReportRollupService,
+    private ssmtReportRollupService: SsmtReportRollupService
 
   ) {
     this.initSummary();
@@ -78,18 +73,14 @@ export class ReportRollupService {
     this.psatReportRollupService.initSummary();
     this.phastReportRollupService.initSummary();
     this.fsatReportRollupService.initSummary();
-
+    this.ssmtReportRollupService.initSummary();
     this.reportAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
 
-    this.ssmtAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
     this.treasureHuntAssessments = new BehaviorSubject<Array<ReportItem>>(Array<ReportItem>());
 
 
 
 
-    this.selectedSsmt = new BehaviorSubject<Array<SsmtCompare>>(new Array<SsmtCompare>());
-    this.ssmtResults = new BehaviorSubject<Array<SsmtResultsData>>(new Array<SsmtResultsData>());
-    this.allSsmtResults = new BehaviorSubject<Array<AllSsmtResultsData>>(new Array<AllSsmtResultsData>());
 
     this.calcsArray = new Array<Calculator>();
     this.selectedCalcs = new BehaviorSubject<Array<Calculator>>(new Array<Calculator>());
@@ -103,12 +94,7 @@ export class ReportRollupService {
     let settings = this.settingsDbService.getByAssessmentId(assessment);
     let tmpSettings = this.checkSettings(settings);
     this.assessmentsArray.push({ assessment: assessment, settings: tmpSettings });
-    if (assessment.ssmt) {
-      this.ssmtArray.push({
-        assessment: assessment,
-        settings: tmpSettings
-      })
-    } else if (assessment.treasureHunt) {
+ if (assessment.treasureHunt) {
       this.treasureHuntArray.push(
         {
           assessment: assessment,
@@ -120,10 +106,7 @@ export class ReportRollupService {
 
   getReportData(directory: Directory) {
     this.assessmentsArray = new Array<ReportItem>();
-    // this.phastArray = new Array<ReportItem>();
-    // this.psatArray = new Array<ReportItem>();
-    // this.fsatArray = new Array<ReportItem>();
-    this.ssmtArray = new Array<ReportItem>();
+    // this.ssmtArray = new Array<ReportItem>();
     this.treasureHuntArray = new Array<ReportItem>();
     this.calcsArray = new Array<Calculator>();
     let selected = directory.assessments.filter((val) => { return val.selected; });
@@ -150,7 +133,8 @@ export class ReportRollupService {
     this.psatReportRollupService.psatAssessments.next(psatArray);
     let fsatArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.fsat != undefined });
     this.fsatReportRollupService.fsatAssessments.next(fsatArray);
-    this.ssmtAssessments.next(this.ssmtArray);
+    let ssmtArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.ssmt != undefined });
+    this.ssmtReportRollupService.ssmtAssessments.next(ssmtArray);
     this.treasureHuntAssessments.next(this.treasureHuntArray);
     this.reportAssessments.next(this.assessmentsArray);
 
@@ -182,86 +166,7 @@ export class ReportRollupService {
       this.pushAssessment(assessment);
     });
   }
-
-
-
-
-
-
-  //used for ssmt summary
-  initSsmtCompare(resultsArr: Array<AllSsmtResultsData>) {
-    let tmpResults: Array<SsmtCompare> = new Array<SsmtCompare>();
-    resultsArr.forEach(result => {
-      let minCost = _.minBy(result.modificationResults, (result) => { return result.operationsOutput.totalOperatingCost; });
-      let modIndex;
-      if (minCost != undefined) {
-        modIndex = _.findIndex(result.modificationResults, (result) => { return result.operationsOutput.totalOperatingCost == minCost.operationsOutput.totalOperatingCost });
-      }
-      let ssmtAssessments = this.ssmtAssessments.value;
-      let assessmentIndex = _.findIndex(ssmtAssessments, (val) => { return val.assessment.id === result.assessmentId; });
-      let item = ssmtAssessments[assessmentIndex];
-      if (result.isBaseline || modIndex == undefined || modIndex == -1) {
-        tmpResults.push({ baseline: item.assessment.ssmt, modification: item.assessment.ssmt, assessmentId: result.assessmentId, selectedIndex: -1, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-      } else {
-        tmpResults.push({ baseline: item.assessment.ssmt, modification: item.assessment.ssmt.modifications[modIndex].ssmt, assessmentId: result.assessmentId, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-      }
-    });
-    this.selectedSsmt.next(tmpResults);
-  }
-
-  updateSelectedSsmt(item: ReportItem, modIndex: number) {
-    let tmpSelected = this.selectedSsmt.value;
-    if (modIndex !== -1) {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.ssmt, modification: item.assessment.ssmt.modifications[modIndex].ssmt, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    } else {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.ssmt, modification: item.assessment.ssmt, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    }
-    this.selectedSsmt.next(tmpSelected);
-  }
-
-  initSsmtResultsArr(ssmtArr: Array<ReportItem>) {
-    let tmpResultsArr = new Array<AllSsmtResultsData>();
-    ssmtArr.forEach(val => {
-      if (val.assessment.ssmt.setupDone) {
-        //get results
-        val.assessment.ssmt.outputData = this.ssmtService.calculateBaselineModel(val.assessment.ssmt, val.settings).outputData;
-        let baselineResults: SSMTOutput = val.assessment.ssmt.outputData;
-        if (val.assessment.ssmt.modifications) {
-          if (val.assessment.ssmt.modifications.length !== 0) {
-            let modResultsArr = new Array<SSMTOutput>();
-            val.assessment.ssmt.modifications.forEach(mod => {
-              mod.ssmt.outputData = this.ssmtService.calculateModificationModel(mod.ssmt, val.settings, baselineResults).outputData;
-              let tmpResults: SSMTOutput = mod.ssmt.outputData;
-              modResultsArr.push(tmpResults);
-            });
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id });
-          } else {
-            let modResultsArr = new Array<SSMTOutput>();
-            modResultsArr.push(baselineResults);
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-          }
-        } else {
-          let modResultsArr = new Array<SSMTOutput>();
-          modResultsArr.push(baselineResults);
-          tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-        }
-      }
-    });
-    this.allSsmtResults.next(tmpResultsArr);
-  }
-
-  getSsmtResultsFromSelected(selectedSsmt: Array<SsmtCompare>) {
-    let tmpResultsArr = new Array<SsmtResultsData>();
-    selectedSsmt.forEach(val => {
-      let baselineResults: SSMTOutput = val.baseline.outputData;
-      let modificationResults: SSMTOutput = val.modification.outputData;
-      tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modificationResults, assessmentId: val.assessmentId, name: val.name, modName: val.modification.name, baseline: val.baseline, modification: val.modification, settings: val.settings });
-    });
-    this.ssmtResults.next(tmpResultsArr);
-  }
-
+  
   //TREASURE HUNT
   initTreasureHuntResultsArray(thuntItems: Array<ReportItem>) {
     let tmpResultsArr: Array<TreasureHuntResultsData> = new Array<TreasureHuntResultsData>();
