+++ src/app/report-rollup/report-summary/report-summary.component.ts
@@ -5,6 +5,7 @@ import { FsatReportRollupService } from '../fsat-report-rollup.service';
 import { PhastReportRollupService } from '../phast-report-rollup.service';
 import { PsatReportRollupService } from '../psat-report-rollup.service';
 import { ReportRollupService } from '../report-rollup.service';
+import { SsmtReportRollupService } from '../ssmt-report-rollup.service';
 
 @Component({
   selector: 'app-report-summary',
@@ -21,11 +22,14 @@ export class ReportSummaryComponent implements OnInit {
   showPsatSummary: boolean;
   showPhastSummary: boolean;
   showFsatSummary: boolean;
+  showSsmtSummary: boolean;
   psatAssessmentsSub: Subscription;
   phastAssessmentsSub: Subscription;
   fsatAssessmentsSub: Subscription;
+  ssmtAssessmentsSub: Subscription;
   constructor(public reportRollupService: ReportRollupService, private psatReportRollupService: PsatReportRollupService,
-    private phastReportRollupService: PhastReportRollupService, private fsatReportRollupService: FsatReportRollupService) { }
+    private phastReportRollupService: PhastReportRollupService, private fsatReportRollupService: FsatReportRollupService,
+    private ssmtReportRollupService: SsmtReportRollupService) { }
 
   ngOnInit() {
     this.psatAssessmentsSub = this.psatReportRollupService.psatAssessments.subscribe(val => {
@@ -39,12 +43,17 @@ export class ReportSummaryComponent implements OnInit {
     this.fsatAssessmentsSub = this.fsatReportRollupService.fsatAssessments.subscribe(val => {
       this.showFsatSummary = val.length != 0;
     });
+
+    this.ssmtAssessmentsSub = this.ssmtReportRollupService.ssmtAssessments.subscribe(val => {
+      this.showSsmtSummary = val.length != 0;
+    })
   }
 
   ngOnDestroy() {
     this.psatAssessmentsSub.unsubscribe();
     this.phastAssessmentsSub.unsubscribe();
     this.fsatAssessmentsSub.unsubscribe();
+    this.ssmtAssessmentsSub.unsubscribe();
   }
 
   showAssessmentModal(assessmentModalType: string) {
