+++ src/app/report-rollup/report-summary/ssmt-summary/ssmt-summary.component.html
@@ -2,7 +2,7 @@
   <div class="card-header">
     <h3>Steam</h3>
     <span>
-      <span>{{reportRollupService.numSsmt}}</span>
+      <span>{{numSsmt}}</span>
     </span>
   </div>
   <div class="card-body">
