+++ src/app/shared/print-options-menu/print-options-menu.component.ts
@@ -8,6 +8,7 @@ import { ReportRollupService } from '../../report-rollup/report-rollup.service';
 import { PsatReportRollupService } from '../../report-rollup/psat-report-rollup.service';
 import { PhastReportRollupService } from '../../report-rollup/phast-report-rollup.service';
 import { FsatReportRollupService } from '../../report-rollup/fsat-report-rollup.service';
+import { SsmtReportRollupService } from '../../report-rollup/ssmt-report-rollup.service';
 @Component({
   selector: 'app-print-options-menu',
   templateUrl: './print-options-menu.component.html',
@@ -27,7 +28,8 @@ export class PrintOptionsMenuComponent implements OnInit {
   showTHReportOptions: boolean = false;
   showWasteWaterOptions: boolean = false;
   constructor(private printOptionsMenuService: PrintOptionsMenuService, private windowRefService: WindowRefService, private reportRollupService: ReportRollupService,
-    private psatReportRollupService: PsatReportRollupService, private phastReportRollupService: PhastReportRollupService, private fsatReportRollupService: FsatReportRollupService) { }
+    private psatReportRollupService: PsatReportRollupService, private phastReportRollupService: PhastReportRollupService, private fsatReportRollupService: FsatReportRollupService,
+    private ssmtReportRollupService: SsmtReportRollupService) { }
 
   ngOnInit() {
     this.setContext();
@@ -63,7 +65,7 @@ export class PrintOptionsMenuComponent implements OnInit {
       this.showPsatReportOptions = (this.psatReportRollupService.psatAssessments.getValue().length != 0);
       this.showFsatReportOptions = (this.fsatReportRollupService.fsatAssessments.getValue().length != 0);
       this.showPhastReportOptions = (this.phastReportRollupService.phastAssessments.getValue().length != 0);
-      this.showSsmtReportOptions = (this.reportRollupService.numSsmt != 0);
+      this.showSsmtReportOptions = (this.ssmtReportRollupService.ssmtAssessments.getValue().length != 0);
       this.showTHReportOptions = (this.reportRollupService.numTreasureHunt != 0);
       // this.showWasteWaterOptions = 
     }
@@ -89,9 +91,6 @@ export class PrintOptionsMenuComponent implements OnInit {
     this.printMenuModal.onHidden.subscribe(() => {
       this.printOptionsMenuService.showPrintView.next(true);
       let tmpPrintBuildTime: number = 2000;
-      // if (this.reportRollupService.numSsmt > 0) {
-      //   tmpPrintBuildTime += (500 * this.reportRollupService.numSsmt);
-      // }
       setTimeout(() => {
         this.print();
       }, tmpPrintBuildTime);
